posix/executor: Fix pipe communication between parent & child
Background: In order to communicate errors from the child to the parent,
a pipe is used. The child writes 2 ints (errno value, length of error
message) followed by N bytes containing the error message.

The parent didn't validate the amount of bytes received, so it sometimes
accessed uninitialized memory.

Specifically, _read_error()'s first read() of length 8 sometimes only
returned 4 bytes, leaving data[1] uninitialized. This caused further
problems since data[1] is the error message length used for the next
memory allocation and read(). This probably only happened when the child
used write_error() (writing 4 + 4 + N bytes) instead of _write_error()
(writing 8 + N bytes).

Either way, read() on a pipe is not necessarily guaranteed to return the
requested amount of bytes, so we have to loop and call it again in order
to get the rest.

A potential danger now is that read_all() waits indefinitely, if the
child process doesn't send the expected/announced amount of bytes.
However, given the current code, I think that can't happen,
plus end-of-file is (still) explicitly handled as early abort.